initial tilemap loader
